+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.4.0 - Nov 25, 2020
+
+-   **Breaking change**: custom logger via static variable `DisposeBag.logger`. It can be set to `null` to disable logging.
+-   Refactor: performance optimization.
+
 ## 1.3.2 - Nov 18, 2020
 
 -   Fix: `The method 'mapIndexed' is defined in multiple extensions for 'Set<dynamic>' and neither is more specific.` when using `collection: 1.15.0-nullsafety.5` package.